Many Changes... See full commit for details. Renamed Database Updated .Equals method in Computer object Cleaned up no-arg constructor for mapping to initialize with empty/default objects Refactored Computer and Server to have a 15 character limit validation on hostname, will now automatically trim incorrent hostnames when calling setter method. Added Extensions.cs with List<NextworkConnection> methods DNSable and DistinctServers Added Copy Constructor to NetworkConnection class
MAJOR refactoring of main program method. Biggest change was how the DNS-ing works, now servers are DNSed and networkconnections are filtered out before building and creating entities. Unresolvable servers will not be processed at all. - Will need to follow up to make sure that non-resolvable but marked as inactive servers are unmapped.

Note: this is a temporary commit, it is not feature full. If a server is non-resolvable but marked as inactive, it will not be unmapped.